+++ Assets/Scripts/MagicAVoxel/Scripts/Vox/VoxAsset.cs
@@ -1,7 +1,9 @@
+using System.Collections.Generic;
 using UnityEngine;
 
 [System.Serializable]
 public class VoxAsset : ScriptableObject
 {
-    public byte[] data;
+    public byte[] rawData;
+    public VoxData data;
 }