+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
+## 0.0.2
+### Add complete support for MagicaVoxel model format
+- Support loading all models in the file
+- Cache mesh and voxel data in asset 
 ## 0.0.1
 ### Initial Commit
 - Add: Initial code